Fix appid for calibration tool
Because the calibration tool did not share the namespace of the main
application you would need to explicitly add ownership of the name
com.github.wwmm.pulseeffects_calibration.

With this simple change the calibration tool appid is under the same
namespace and granted dbus acces by default in the flatpak sandbox.