Initial meson implementation
Add support for the meson build system and restructure how data files
is stored in the repository.

Meson has the advantage over setuptools that it will automatically
install all files to the correct locations for the respective operating
system or distribution. No more manual copying of files post build.

Supporting AppStream appdata, including translations is very complicated
without a native build system. By using Meson it is possible to
translate both the application, desktop entry, appdata and even the
gschema files, in a single PO file. This means a much simpler workflow
for translators.

The desktop entry file is renamed according to the freedesktop, desktop
entry specification, (using the "reverse DNS" convention). This is
required to support certain tags in AppStream appdata and remves the
need to rename it at build time (Flatpak).