Improve support for buildtime options
Use standard build time options everywhere in favor of static values
like 'bin'. These options are now checked once in the main build file.

There is a limitation with the datadir build option. If used in
combination with the DESTDIR env var, the value of datadir will not
be respected. by the post install script. It simply can't because there
are no env vars exposed for the datadir option.

However most of the time 'share' (default) will be the datadir. So it
should not be an issue when the FHS directory structure is followed to
to some degree.