Properly initialize gettext
Translation support only worked if the required files were installed to
the default locale directory (in python it is /usr/share/locale).
Because of this it did not work at all with Flatpak, because that uses a
/app prefix.

By properly initializing gettext this issue is resolved and translations
will work independent of how the application is installed.

The gettext manual mentions that the recommended way of providing
"PACKAGE" (app name) and "LOCALEDIR" is via either the config.h file
(not possible with python), or the Makefile.

This closely follows those recommendations, by using Meson and treating
the initial python script like a config file.