Properly initialize gettext
Without properly initializing gettext globally, translations will not
work in a prefix other than `/usr`. By properly initializing gettext
at an early stage translations will work irrespective of where the
localisation data is installed.

The gettext manual mentions that the recommended way of providing
"PACKAGE" and "LOCALEDIR" is via either the config.h file
or the Makefile. Neither is available, but Meson does provide the
standad `--libdir` option.

https://www.gnu.org/software/gettext/manual/html_node/Triggering.html#Triggering

By treating the pulseeffects launcher script as a configuration file it
is possible to use the values from the build system.