Inform translators not to translate icon name
It seems to be a common issue with Gnome apps, that for some reason the
icon name string is made translatable.

This temporary solution seems to be widely accepted. It is not visible
in POEdit. Maybe it is in other editors, maybe not. At least it will
serve as a reminder during a pull request review.

DO NOT accept pull requests that fail to notice this.