Add meson version check
If nothing else this documents what version of Meson is required to
build/install PulseEffects.

Unfortunately since version 0.40.0 brings the change that removes the
requirement to define a language in project() and that is the very same
place the meson version check happens Meson will still report:

"Not enough arguments to project(). Needs at least the project name and
one language"