Update appdata screenshots
It is good to note that screenshots should preferably be 16:9 according
to the AppStream specification.

Since PulseEffects windows are note 16:9 by default, it is a bit of a
hassle to meet this requirement. Currently the workaround used is using
`xdotool` to set a window size of "1244 700" for the main PulseEffects
window and a size of "886 521" for Calibration.