Update pot and po files
Update all localizations against current template. Done by running:
ninja -C $builddir PulseEffects-update-po`

This is the equivalent to each translator updating their file against
the new template. Any strings that are already translated will
automatically be reused.

It is probably a good idea to to this at least prior to a release.