Fix issues with documentation
The recent introduction of debian build files also modified the
documentation in less than ideal ways.

This fix solves this by staying consistent with previous changes that
moved distribution specific instructions to the wiki. Specifically a
wiki page for community maintained distribution packages is introduced
and linked to.

Build instructions is removed altogether since it has long since had a
dedicated wiki page. Some instructions were incomplete anyhow or located
in the wrong section.

Showing incomplete build instructions at the top is bad, because it
decreases the likelihood of a user reading the complete wiki page, thus
missing information. This in turn leads to an increase in support issues.