Emphasize that PulseEffects is equalizer
It might be not convenient for the ordinary user to find pulseeffects application from KDE/Gnome. I suggest we emphasize that PulseEffects is equalizer mostly. Eventually we will benefit once PulseEffects becomes a part of default set of applications installed with major distributions. Users will try to type in "equalizer" and will be able to find PulseEffects because "PulseEffects" is too technical (relates to pulseaudio), not everyone will know it / remember it.