Add OARS 1.0 data to AppStream file
The "Open Age Ratings Service" free and commercial software to
inform users about data collection, and content that may be violent or
otherwise should be age restricted.

Basically it is just a way to tag content in so that users (and parents)
can be informed prior to installation and potential exposure.

While not that relevant to PulseEffects, given all values are set to
"none", it is still valuable to show this. That and it is now listed as
a submission requirement at Flathub.

Version 1.0 is used for now, in favor of 1.1 which only works with
GNOME Software >= 3.27.3. Other clients are unknown, while 1.0 is known
to be supported.