Update Flatpak install instructions
The LSP plugin collection uses a very restrictive license, that may not
even be considered open source (according to the OSI definition). On top
of this the recently added Expander plugin is currently proprietary.

In order to provide both the Expander and Delay plugins an extension
(com.github.wwmm.pulseeffects.plugin.lsp-plugins) has been added to
flathub. This solves the incompatibility with the GPL and makes it
optional if users don't want to install proprietary plugins.

The new extension is provided for x86_64 and i386. ARM is not supported,
because upstream does not provide any binaries.