Fix AppStream validation errors ("Equalizer")
Apparently appstream-glib (`appstream-util validate`) does not like
captions being less than 10 characters. Instead a short summary is used.

In the event of the screenshots being updated and/or exchanged for
screenshots that feature other areas of the app, one should consider
migrating to short 10-50 character summaries instead of a single word.