Add script that adds new releases to appdata
This script automates the additions to release data in the AppStream
metainfo (appdata) file. It eliminates many possible human errors and
ensures the dates are consistent and always based upon the committed timezone.

The process of adding XML data is based on XML Style Sheets. This is
much safer and more reliable than simply using a tool like `sed`.

If the appdata already contains a release that matches the last tagged
version, no action will be taken.

Together all of this means it should be practically impossible for this
tool to accidentally overwrite any other data. Of course, it does
not hurt to first stage or commit any other changes.