Fix add-appdata-release utility
The initial script had a fundamental issue; the application version and
release dates were based on git tags. Since this script needs to be run
before a release is tagged, that tag will never include an up-to-date
appdata file.

Versions are now parsed from the main meson.build file. It works well
but there is one limitation, the order of the key/value pairs "version"
and "meson_version" matters. If the order of those keys changes this
script would also require adjustments.

Dates are now provided by the `date` command, using the UTC setting.
This offers consistency since it also is not affected by day light
savings.

An added benefit of these changes is that less external commands are
now used.