Remove appdata release descriptions
This data is currently impractical, and even less practical to translate.

Since PulseEffects does very rapid releases, there is little time to
write proper, easy to read, release descriptions.

Another related problem is translations. There is no window of time for
translators to localize this data. This data could be marked so that it
is not translatable, but some meson or xgettext issues seem to prevent
this from working.

If this data is wanted in the future, simply revert this commit.

Fixes: #322