throw exception when a required plugin can't be found
This can happen if e.g. a user forgot to install required gstreamer
plugins or a package maintainer forgot to add dependencies.
If a required plugin is missing, pulseeffects is effectively broken and
should crash hard, so that e.g. package maintainers get notified and can
fix the package dependencies. The user should be clearly alerted that
something is missing and the installation is somehow botched.