Multiband Gate: Make clear Reduction is Max Reduction
The "Reduction" control sounds like it is used to apply a specified
fixed reduction. That is not correct. It actually acts as a limit
on how much reduction will be applied by the value of Ratio. This
rename will make in less likely for audio effects newbies (like
myself) to be frustrated and confused when trying to figure out how
this plugin works.

See the documentation for the non-multiband Gate at:
https://calf-studio-gear.org/doc/Gate.html