spectrum: precompute Hann window
Computing the factor applied to each sample is the most expensive
operation in the whole process function, by a large margin. Avoid that
by precomputing the 8192 factors.

A memory size optimization could be to compute half of it, as it is
symmetrical. It would make the code more complex for small gains; we
chase the low hanging fruits here.

Here is a before/after of performances, acquired by running pw-profiler
during 15s, with a 512 samples quantum:

         Before      After
calls      1545       1536
avg (µs)    183.1       65.6
stdev        29.0       15.7
min          43         16
5th perc    131         40
25th perc   170         57
median      182         64
75th perc   200         76
95th perc   226         94
max         365        114

Signed-off-by: Théo Lebrun <theo.lebrun@bootlin.com>