spectrum: replace input buffer std::deque by std::vector
deque_in_mono is a mono input buffer. Its size invariants are not
straight forward to understand (see "spectrum: remove impossible
condition" commit). Replace it by a plain vector storing a sample
history. It is of size n_bands, ie the amount of samples required to
compute a DFT.

This avoids:
 - any potential allocation in ::setup() or ::process(),
 - the complex implementations of deque::push_back() and ::pop_front(),
 - cleanup required at ::process()'s end.

We empty the history at ::setup(). It could be safe to keep old data but
it probably does not make much sense: we might have been re-linked with
another source.

We handle properly the edge case of `n_samples > n_bands`.
This setup is highly unlikely.

The standard case is `n_samples < n_bands`. Here, we (1) drop the oldest
quantum by shifting all the samples and (2) fill the buffer using new
data. This implies we have nothing to do for cleanup at the end of
process().

Here is a before/after of performances, acquired by running pw-profiler
during 15s, with a 512 samples quantum:

         Before      After
calls      1536       1539
avg (µs)     65.6       50.7
stdev        15.7       13.8
min          16         13
5th perc     40         30
25th perc    57         43
median       64         51
75th perc    76         58
95th perc    94         75
max         114        106

Signed-off-by: Théo Lebrun <theo.lebrun@bootlin.com>