spectrum: avoid data_mutex locking in process()
data_mutex is not locked in PluginBase. It is locked inside spectrum:
 - in the process() callback,
 - on settings change callback (for bypass),
 - in the destructor.

Now that bypass is atomic, no need to protect/lock accesses to it.
The destructor starts by disconnecting the filter from PipeWire so its
process() function cannot be invoked.

We can therefore remove this potential sleeping mechanism from our
realtime thread.

Signed-off-by: Théo Lebrun <theo.lebrun@bootlin.com>