Add back support for devel build This has a seperate icon and display name. The names of a lot of internal things are still the same.
The main idea here is to allow flatpak to fully work with a seperate application id. So things like the desktop file and metainfo files need to be adjusted, as well as some internal ids so portals can work. But otherwise a lot of references to things like "easyeffects" should possible to leave the same.